requirements.txt
PyOpenGL
PyOpenGL_accelerate
pyautogui
glfw
imgui
keyboard
pywin32